hwcomposer: correct the type of buffers in structure _VASurfaceAttributeTPI
BZ: 172915

correct the type of buffers in structure _VASurfaceAttributeTPI

+++ ips/common/RotationBufferProvider.cpp
@@ -145,7 +145,7 @@ bool RotationBufferProvider::createVaSurface(VideoPayloadBuffer *payload, int tr
     VASurfaceAttributeTPI attribTpi;
     VASurfaceAttributeTPI *vaSurfaceAttrib = &attribTpi;
     int stride;
-    unsigned int buffers;
+    unsigned long buffers;
     VASurfaceID *surface;
     int width = 0, height = 0, bufferHeight = 0;
